refactored blank history data into util
