+++ routes/tools_routes.js
@@ -5,6 +5,7 @@ const express = require('express');
 
 const carddb = require('../serverjs/cards');
 const cardutil = require('../dist/utils/Card.js');
+const getBlankCardHistory = require('../src/utils/BlankCardHistory.js');
 const { filterUses, makeFilter, filterCardsDetails } = require('../dist/filtering/FilterCards');
 const generateMeta = require('../serverjs/meta.js');
 const util = require('../serverjs/util.js');
@@ -301,65 +302,7 @@ router.get('/card/:id', async (req, res) => {
     let data = await CardHistory.findOne({ oracleId: card.oracle_id });
     // id is valid but has no matching history
     if (!data) {
-      const cardVersions = carddb.getVersionsByOracleId(card.oracle_id);
-
-      data = {
-        current: {
-          rating: null,
-          elo: card.elo,
-          picks: 0,
-          total: [0, 0],
-          size180: [0, 0],
-          size360: [0, 0],
-          size450: [0, 0],
-          size540: [0, 0],
-          size720: [0, 0],
-          pauper: [0, 0],
-          legacy: [0, 0],
-          modern: [0, 0],
-          standard: [0, 0],
-          vintage: [0, 0],
-          cubes: 0,
-          prices: cardVersions.map((cardId) => {
-            return { ...carddb.cardFromId(cardId), version: carddb.cardFromId(cardId)._id };
-          }),
-        },
-        cubedWith: {
-          synergistic: [],
-          top: [],
-          creatures: [],
-          spells: [],
-          other: [],
-        },
-        versions: cardVersions,
-        cubes: [],
-        history: [
-          {
-            data: {
-              size180: [0, 0],
-              size360: [0, 0],
-              size450: [0, 0],
-              size540: [0, 0],
-              size720: [0, 0],
-              pauper: [0, 0],
-              legacy: [0, 0],
-              modern: [0, 0],
-              standard: [0, 0],
-              vintage: [0, 0],
-              total: [0, 0],
-              rating: null,
-              elo: null,
-              picks: 0,
-              cubes: 0,
-              prices: [],
-            },
-            date: '',
-          },
-        ],
-        cardName: card.name,
-        oracleId: card.oracle_id,
-        __v: 0,
-      };
+      data = getBlankCardHistory(id);
     }
     const related = {};
